services/ocr-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.4
pillow>=10.0.0
reportlab>=4.0.0
